Add check user by room admin (#2)

+++ messenger-service/app/api/routers/rooms.py
@@ -2,6 +2,7 @@ from uuid import uuid4, UUID
 
 from app.repositories.rooms import RoomsRepository
 from app.schemes.messengers import Room, RoomInDB, RoomMembers, User
+from app.tools.security import check_user_is_room_admin
 from fastapi import APIRouter, Depends, status, Request
 from loguru import logger
 
@@ -26,6 +27,8 @@ async def create_room(request: Request, payload: Room,
 @router.delete('/{room_id}', status_code=status.HTTP_200_OK)
 async def delete_room(request: Request, room_id: UUID,
                       rooms: RoomsRepository = Depends()):
+    check_user_is_room_admin(room_id, request.user.id, rooms)
+
     await rooms.delete(room_id)
     logger.info(f"Successfully deleted room with id {room_id} "
                 f"by username {request.user.first_name}")
@@ -35,6 +38,8 @@ async def delete_room(request: Request, room_id: UUID,
               status_code=status.HTTP_200_OK)
 async def update_room(request: Request, room: RoomInDB,
                       rooms: RoomsRepository = Depends()):
+    check_user_is_room_admin(room.id, request.user.id, rooms)
+
     result = await rooms.update(room)
     logger.info(f"Successfully update room with id {room.id} "
                 f"by username {request.user.first_name}")
@@ -52,8 +57,10 @@ async def find_room_by_id(room_id: UUID, rooms: RoomsRepository = Depends()):
 
 @router.get("/members/{room_id}", response_model=RoomMembers,
             status_code=status.HTTP_200_OK)
-async def find_all_members_by_room_id(room_id: UUID,
+async def find_all_members_by_room_id(request: Request, room_id: UUID,
                                       rooms: RoomsRepository = Depends()):
+    check_user_is_room_admin(room_id, request.user.id, rooms)
+
     result = await rooms.find_all_members(room_id)
 
     return RoomMembers(
@@ -65,6 +72,8 @@ async def find_all_members_by_room_id(room_id: UUID,
 @router.post('/members', status_code=status.HTTP_201_CREATED)
 async def add_members_to_room(request: Request, room_members: RoomMembers,
                               rooms: RoomsRepository = Depends()):
+    check_user_is_room_admin(room_members.room_id, request.user.id, rooms)
+
     await rooms.add_members(room_members)
     logger.info(f"Successfully added {len(room_members.members)} members "
                 f"to room with id {room_members.room_id} "
@@ -74,6 +83,8 @@ async def add_members_to_room(request: Request, room_members: RoomMembers,
 @router.delete('/members', status_code=status.HTTP_200_OK)
 async def delete_members_from_room(request: Request, room_members: RoomMembers,
                                    rooms: RoomsRepository = Depends()):
+    check_user_is_room_admin(room_members.room_id, request.user.id, rooms)
+
     await rooms.delete_members(room_members)
     logger.info(f"Successfully deleted {len(room_members.members)} members "
                 f"from room with id {room_members.room_id} "
